chore(@clayui/list): update List component stories
